[bug 1203173] Move hb analyzer views and add tests
* move heartbeat analyzer views to fjord/heartbeat/
* add basic tests for the heartbeat views
